functions added to utilities so it can be called from other py files, and other minor fixes
